mv get chain info function to subnet
